+++ docs/conf.py
@@ -8,7 +8,7 @@ sys.path.insert(0, os.path.abspath("../src"))
 project = "microfactual"
 copyright = "2025, Simeon Hebrew, Lawrence Adu-Gyamfi"
 author = "Simeon Hebrew, Lawrence Adu-Gyamfi"
-release = "0.1.0"
+release = "0.2.0"
 
 extensions = [
     "sphinx.ext.autodoc",